Foundations for Calendar set. Refactored some old confusing code and package reference (medmemory -> memorymed) for consistency. Added Fundamental SharedPreferences support, available to all activities using MainActivity.PREF_NAME reference.
New Classes:
Calendar
CalenddarEventActivity (stub)
CalendarSelectionDialog

New Resources:
calendar.xml
