+++ SE4351_class_project-master/MedMemory/.project
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <projectDescription>
-	<name>MedMemory</name>
+	<name>MemoryMed</name>
 	<comment></comment>
 	<projects>
 	</projects>
