+++ SE4351_class_project-master/MedMemory/src/com/example/memorymed/Pharm.java
@@ -1,4 +1,4 @@
-package com.example.medmemory;
+package com.example.memorymed;
 
 import android.app.Activity;
 import android.os.Bundle;
