+++ SE4351_class_project-master/MedMemory/src/com/example/memorymed/model/Medfriend.java
@@ -1,4 +1,4 @@
-package com.example.medmemory.model;
+package com.example.memorymed.model;
 
 /**
  * Created by Cameron on 2/5/2015.
